+++ smart-contracts/cargo-concordium/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
 ## Unreleased changes
+- Clarify that energy units used by `cargo-concordium` are "interpreter energy"
+  and not the same as NRG.
 - Allow the user to only specify the necessary fields in the JSON context files
   - Also allow the `--context` parameter to be omitted, for when no context is needed